Develop

+++ plugins/action/license_device_license_summary_info.py
@@ -2,8 +2,7 @@
 # -*- coding: utf-8 -*-
 
 # Copyright (c) 2021, Cisco Systems
-# GNU General Public License v3.0+ (see LICENSE or
-# https://www.gnu.org/licenses/gpl-3.0.txt)
+# GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
 
@@ -34,7 +33,7 @@ argument_spec.update(
         sort_by=dict(type="str"),
         dna_level=dict(type="str"),
         device_type=dict(type="str"),
-        limit=dict(type="float"),
+        limit=dict(type="int"),
         registration_status=dict(type="str"),
         virtual_account_name=dict(type="str"),
         smart_account_id=dict(type="float"),
