+++ plugins/action/site_health_summaries_trend_analytics_info.py
@@ -2,8 +2,7 @@
 # -*- coding: utf-8 -*-
 
 # Copyright (c) 2021, Cisco Systems
-# GNU General Public License v3.0+ (see LICENSE or
-# https://www.gnu.org/licenses/gpl-3.0.txt)
+# GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
 
@@ -36,8 +35,8 @@ argument_spec.update(
         siteType=dict(type="str"),
         id=dict(type="str"),
         trendInterval=dict(type="str"),
-        limit=dict(type="float"),
-        offset=dict(type="float"),
+        limit=dict(type="int"),
+        offset=dict(type="int"),
         timeSortOrder=dict(type="str"),
         attribute=dict(type="str"),
         taskId=dict(type="str"),
